Convert build system to GNU configure & build
This patch adds a standard configure script and converts the Makefile to
automake.  This adresses PR #30, but in a more generic way, and also
eliminates the hard coded CC variable, enabling building with clang or
even a cross compiler.

Note: this is more a proof of concept and does not include the man page
      generation.  Another pull request addresses that.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>