Rearrange and simplify text a bit, also put examples early
Most users want to see the examples as early as possible, so we place
them directly after the new (simplified) introduction.  The examples
have been rewritten to separate commands from options and grouping
the command with its value, where applicable.

The rest of the document has just been auto-filled to fit to < 72
columns and sections now use === and ---- for headings to make the
README more accessible from a terminal.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>